requirements.txt
requests
beautifulsoup4
awesome-slugify
